+++ extensions/toggl-track/src/components/CreateTimeEntryForm.tsx
@@ -1,15 +1,46 @@
-import { useNavigation, Form, ActionPanel, Action, Icon, showToast, Toast, clearSearchBar } from "@raycast/api";
-import { createTimeEntry, Project, Task } from "../api";
-import { useTimeEntryContext } from "../context/TimeEntryContext";
 import { useMemo, useState } from "react";
+import { useNavigation, Form, ActionPanel, Action, Icon, showToast, Toast, clearSearchBar } from "@raycast/api";
+import { useCachedState } from "@raycast/utils";
+import { createTimeEntry, Project, Task } from "../api";
+import { useMe, useWorkspaces, useClients, useTags, useTasks, useEffectWithCachedDeps } from "../hooks";
+import { createProjectGroups, ProjectGroup } from "../helpers/createProjectGroups";
 
-function CreateTimeEntryForm({ project, description }: { project?: Project; description?: string }) {
+interface CreateTimeEntryFormParams {
+  isLoading: boolean;
+  projects: Project[];
+  revalidateRunningTimeEntry: () => void;
+  project?: Project;
+  description?: string;
+}
+
+function CreateTimeEntryForm({
+  isLoading,
+  projects,
+  revalidateRunningTimeEntry,
+  project,
+  description,
+}: CreateTimeEntryFormParams) {
   const navigation = useNavigation();
-  const { me, isLoading, projects, tags, tasks, projectGroups, revalidateRunningTimeEntry } = useTimeEntryContext();
+  const { me, isLoadingMe } = useMe();
+  const { workspaces, isLoadingWorkspaces } = useWorkspaces();
+  const { clients, isLoadingClients } = useClients();
+  const { tags, isLoadingTags } = useTags();
+  const { tasks, isLoadingTasks } = useTasks();
   const [selectedProject, setSelectedProject] = useState<Project | undefined>(project);
   const [selectedTags, setSelectedTags] = useState<string[]>([]);
   const [selectedTask, setSelectedTask] = useState<Task | undefined>();
   const [billable, setBillable] = useState<boolean>(false);
+
+  const [projectGroups, setProjectGroups] = useCachedState<ProjectGroup[]>("projectGroups", []);
+
+  useEffectWithCachedDeps(
+    () => {
+      const projectGroups = createProjectGroups(projects, workspaces, clients);
+      setProjectGroups(projectGroups);
+    },
+    [projects, workspaces, clients],
+    toggleArrayIsEqual,
+  );
 
   async function handleSubmit(values: { description: string }) {
     const workspaceId = selectedProject?.workspace_id || me?.default_workspace_id;
@@ -60,7 +91,7 @@ function CreateTimeEntryForm({ project, description }: { project?: Project; desc
 
   return (
     <Form
-      isLoading={isLoading}
+      isLoading={isLoading || isLoadingMe || isLoadingWorkspaces || isLoadingClients || isLoadingTags || isLoadingTasks}
       actions={
         <ActionPanel>
           <Action.SubmitForm title="Create Time Entry" onSubmit={handleSubmit} />
@@ -117,3 +148,7 @@ function CreateTimeEntryForm({ project, description }: { project?: Project; desc
 }
 
 export default CreateTimeEntryForm;
+
+function toggleArrayIsEqual<T extends { id: number }[]>(original: T, updated: T) {
+  return original.length === updated.length && original.every((item, i) => item.id == updated[i].id);
+}
